+++ internal/config/default.go
@@ -45,7 +45,7 @@ func Default() Config {
 			SessionTTL:            Duration(8 * time.Hour),
 			IdleTimeout:           Duration(30 * time.Minute),
 			LoginRPM:              5,
-			MFAPolicy:             "optional",
+			MFAPolicy:             AdminMFAPolicyOptional,
 			DeveloperWorkbench:    "enabled",
 			ReauthElevationWindow: durationPointer(defaultReauthElevationWindow),
 			ModelCapabilityDetection: ModelCapabilityDetection{